fix(rbac): move permission-policy.ts to policies directory
