+++ plugins/rbac-backend/src/file-permissions/csv-file-watcher.test.ts
@@ -22,6 +22,7 @@ import {
   RoleMetadataDao,
   RoleMetadataStorage,
 } from '../database/role-metadata';
+import { ADMIN_ROLE_AUTHOR } from '../policies/permission-policy';
 import { BackstageRoleManager } from '../role-manager/role-manager';
 import { EnforcerDelegate } from '../service/enforcer-delegate';
 import { MODEL } from '../service/permission-model';
